chore: Remove namespaces
